changed doubles to floats

+++ include/neural_network.hpp
@@ -30,9 +30,9 @@ public:
 
 
 	// Feed forward the network to evaluate the checker board.
-	double network_evaluate(std::vector<double> & board_input)
+	float network_evaluate(std::vector<float> & board_input)
 	{
-		double evaluation_value = 0.0;
+		float evaluation_value = 0.0;
 
 		int network_input_size = _network[0].size();
 		int input_layer = 0;
@@ -82,7 +82,7 @@ private:
 	}
 
 
-	double sigmoid(double input)
+	float sigmoid(float input)
 	{
 		return input/(1.0 + abs(input));
 	}
